+++ include/minizinc/ast.hpp
@@ -272,8 +272,12 @@ inline long long int Id::idn() const {
   if (hasStr()) {
     return -1;
   }
-  long long int i = reinterpret_cast<ptrdiff_t>(_vOrIdn.idn) & ~static_cast<ptrdiff_t>(1);
-  return i >> 1;
+  auto i_u = reinterpret_cast<mzn_uintptr_t>(_vOrIdn.idn) & ~static_cast<mzn_uintptr_t>(1);
+  if (static_cast<mzn_intptr_t>(i_u) < 0) {
+    auto ret = static_cast<mzn_intptr_t>(~(~i_u >> 1));
+    return static_cast<long long int>(ret);
+  }
+  return static_cast<long long int>(i_u >> 1);
 }
 
 inline TIId::TIId(const Location& loc, const std::string& v)
@@ -480,7 +484,7 @@ inline UnOp::UnOp(const Location& loc, UnOpType op, Expression* e)
 }
 
 inline bool Call::hasId() const {
-  return (reinterpret_cast<ptrdiff_t>(_uId.decl) & static_cast<ptrdiff_t>(1)) == 0;
+  return (reinterpret_cast<mzn_uintptr_t>(_uId.decl) & static_cast<mzn_uintptr_t>(1)) == 0;
 }
 
 inline ASTString Call::id() const { return hasId() ? _uId.id : decl()->id(); }
@@ -493,14 +497,14 @@ inline void Call::id(const ASTString& i) {
 
 inline FunctionI* Call::decl() const {
   return hasId() ? nullptr
-                 : reinterpret_cast<FunctionI*>(reinterpret_cast<ptrdiff_t>(_uId.decl) &
-                                                ~static_cast<ptrdiff_t>(1));
+                 : reinterpret_cast<FunctionI*>(reinterpret_cast<mzn_uintptr_t>(_uId.decl) &
+                                                ~static_cast<mzn_uintptr_t>(1));
 }
 
 inline void Call::decl(FunctionI* f) {
   assert(f != nullptr);
-  _uId.decl =
-      reinterpret_cast<FunctionI*>(reinterpret_cast<ptrdiff_t>(f) | static_cast<ptrdiff_t>(1));
+  _uId.decl = reinterpret_cast<FunctionI*>(reinterpret_cast<mzn_uintptr_t>(f) |
+                                           static_cast<mzn_uintptr_t>(1));
 }
 
 inline unsigned int Call::argCount() const {